feat: Add support for filtering entities by name
This commit adds the ability to filter entities by name when retrieving entities for a specific device. The `get_entities_for_device` function now accepts an optional `entity_names` parameter, which is a list of entity names or IDs to include in the result. If no `entity_names` are provided, all entities associated with the device will be returned.

Additionally, the configuration flow now includes an optional field called "entity_names" where users can specify a comma-separated list of entity names to pull.

The following files were modified:
- __init__.py
- config_flow.py

The following files were added:
- camera.py
- services.py
- websocket.py

The following translation file was modified:
- en.json